requirements.txt
# Minimal requirements for PartsPro API + V4 Interface
flask==2.3.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
gunicorn==21.2.0

# PDF processing
pymupdf==1.23.6
Pillow==10.1.0

# Web scraping and API integration
google-search-results==2.4.2
requests==2.31.0

# OpenAI integration
openai==1.3.0

# Security
cryptography==41.0.4

# Utilities
pydantic==2.5.1

# Database
psycopg2-binary==2.9.7